% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_load_archive.R
\name{tar_load_archive}
\alias{tar_load_archive}
\title{Load a target's value from archive storage}
\usage{
tar_load_archive(
  names,
  package,
  pipeline,
  branches = NULL,
  meta = NULL,
  strict = TRUE,
  silent = FALSE,
  envir = parent.frame(),
  store = targets::tar_config_get("store")
)
}
\arguments{
\item{names}{Names of the targets to load.
\code{\link[targets:tar_load]{tar_load()}} uses non-standard evaluation in the \code{names} argument
(example: \code{tar_load(names = everything())}), whereas \code{\link[targets:tar_load_raw]{tar_load_raw()}}
uses standard evaluation for \code{names}
(example: \code{tar_load_raw(names = quote(everything()))}).

The object supplied to \code{names} should be a
\code{tidyselect} expression like \code{\link[targets:any_of]{any_of()}} or \code{\link[targets:starts_with]{starts_with()}}
from \code{tidyselect} itself, or \code{\link[targets:tar_described_as]{tar_described_as()}} to select target names
based on their descriptions.}

\item{package}{A scalar character of the package name.}

\item{pipeline}{A scalar character of the pipeline name.}

\item{branches}{Integer of indices of the branches to load
for any targets that are patterns.}

\item{meta}{Data frame of target metadata from \code{\link[targets:tar_meta]{tar_meta()}}.}

\item{strict}{Logical of length 1, whether to error out
if one of the selected targets is in the metadata
but cannot be loaded.
Set to \code{FALSE} to just load the targets in the metadata
that can be loaded and skip the others.}

\item{silent}{Logical of length 1. Only relevant when
\code{strict} is \code{FALSE}. If \code{silent} is \code{FALSE}
and \code{strict} is \code{FALSE}, then a message will be printed
if a target is in the metadata but cannot be loaded.
However, load failures
will not stop other targets from being loaded.}

\item{envir}{R environment in which to load target return values.}

\item{store}{Character of length 1, directory path to the data store
of the pipeline.}
}
\value{
Nothing.
}
\description{
tarchives version of \code{\link[targets:tar_load]{targets::tar_load()}}. Reads one or more targets from
an archived pipeline's store and assigns them into an environment.
}
\examples{
\donttest{
withr::with_envvar(
  c(R_USER_CACHE_DIR = tempfile()),
  {
    tar_make_archive(package = "tarchives", pipeline = "example-model")
    tar_load_archive(model, package = "tarchives", pipeline = "example-model")
    model
  }
)
}

}
