% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_manifest_archive.R
\name{tar_manifest_archive}
\alias{tar_manifest_archive}
\title{List the targets of an archived pipeline}
\usage{
tar_manifest_archive(
  package,
  pipeline,
  names = NULL,
  fields = NULL,
  drop_missing = TRUE,
  callr_function = callr::r,
  callr_arguments = targets::tar_callr_args_default(callr_function),
  envir = parent.frame(),
  script = targets::tar_config_get("script")
)
}
\arguments{
\item{package}{A scalar character of the package name.}

\item{pipeline}{A scalar character of the pipeline name.}

\item{names}{Names of the targets to show. Set to \code{NULL} to
show all the targets (default). Otherwise,
the object supplied to \code{names} should be a
\code{tidyselect} expression like \code{\link[targets:any_of]{any_of()}} or \code{\link[targets:starts_with]{starts_with()}}
from \code{tidyselect} itself, or \code{\link[targets:tar_described_as]{tar_described_as()}} to select target names
based on their descriptions.}

\item{fields}{Names of the fields, or columns, to show. Set to \code{NULL} to
show all the fields (default). Otherwise, the value of \code{fields} should be
a \code{tidyselect} expression like \code{\link[targets:starts_with]{starts_with()}} to select the columns
to show in the output.
Possible fields are below. All of them can be set in \code{\link[targets:tar_target]{tar_target()}},
\code{\link[targets:tar_target_raw]{tar_target_raw()}}, or \code{\link[targets:tar_option_set]{tar_option_set()}}.
\itemize{
\item \code{name}: Name of the target.
\item \code{command}: the R command that runs when the target runs.
\item \code{description}: custom free-form text description of the target,
if available.
\item \code{pattern}: branching pattern of the target, if applicable.
\item \code{format}: Storage format.
\item \code{repository}: Storage repository.
\item \code{iteration}: Iteration mode for branching.
\item \code{error}: Error mode, what to do when the target fails.
\item \code{memory}: Memory mode, when to keep targets in memory.
\item \code{storage}: Storage mode for high-performance computing scenarios.
\item \code{retrieval}: Retrieval mode for high-performance computing scenarios.
\item \code{deployment}: Where/whether to deploy the target in high-performance
computing scenarios.
\item \code{priority}: Numeric of length 1 between 0 and 1. Controls which
targets get deployed first when multiple competing targets are ready
simultaneously. Targets with priorities closer to 1 get dispatched
earlier (and polled earlier in \code{\link[targets:tar_make_future]{tar_make_future()}}).
\item \code{resources}: A list of target-specific resource requirements for
\code{\link[targets:tar_make_future]{tar_make_future()}}.
\item \code{cue_mode}: Cue mode from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_depend}: Depend cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_expr}: Command cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_file}: File cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_format}: Format cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_repository}: Repository cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{cue_iteration}: Iteration cue from \code{\link[targets:tar_cue]{tar_cue()}}.
\item \code{packages}: List columns of packages loaded before running the target.
\item \code{library}: List column of library paths to load the packages.
}}

\item{drop_missing}{Logical of length 1, whether to automatically omit
empty columns and columns with all missing values.}

\item{callr_function}{A function from \code{callr} to start a fresh clean R
process to do the work. Set to \code{NULL} to run in the current session
instead of an external process (but restart your R session just before
you do in order to clear debris out of the global environment).
\code{callr_function} needs to be \code{NULL} for interactive debugging,
e.g. \code{tar_option_set(debug = "your_target")}.
However, \code{callr_function} should not be \code{NULL} for serious
reproducible work.}

\item{callr_arguments}{A list of arguments to \code{callr_function}.}

\item{envir}{An environment, where to run the target R script
(default: \verb{_targets.R}) if \code{callr_function} is \code{NULL}.
Ignored if \code{callr_function} is anything other than \code{NULL}.
\code{callr_function} should only be \code{NULL} for debugging and
testing purposes, not for serious runs of a pipeline, etc.

The \code{envir} argument of \code{\link[targets:tar_make]{tar_make()}} and related
functions always overrides
the current value of \code{tar_option_get("envir")} in the current R session
just before running the target script file,
so whenever you need to set an alternative \code{envir}, you should always set
it with \code{tar_option_set()} from within the target script file.
In other words, if you call \code{tar_option_set(envir = envir1)} in an
interactive session and then
\code{tar_make(envir = envir2, callr_function = NULL)},
then \code{envir2} will be used.}

\item{script}{Character of length 1, path to the
target script file. Defaults to \code{tar_config_get("script")},
which in turn defaults to \verb{_targets.R}. When you set
this argument, the value of \code{tar_config_get("script")}
is temporarily changed for the current function call.
See \code{\link[targets:tar_script]{tar_script()}},
\code{\link[targets:tar_config_get]{tar_config_get()}}, and \code{\link[targets:tar_config_set]{tar_config_set()}} for details
about the target script file and how to set it
persistently for a project.}
}
\value{
A data frame of information about the targets in the pipeline.
Rows appear in topological order (the order they will run
without any influence from parallel computing or priorities).
}
\description{
tarchives version of \code{\link[targets:tar_manifest]{targets::tar_manifest()}}. Returns a data frame of the
targets defined in an archived pipeline's target script file.
}
\examples{
\donttest{
withr::with_envvar(
  c(R_USER_CACHE_DIR = tempfile()),
  tar_manifest_archive(package = "tarchives", pipeline = "example-model")
)
}

}
\seealso{
\code{\link[=tar_archive_pipelines]{tar_archive_pipelines()}} to list the pipelines in a package.
}
