% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_archive.R
\name{tar_archive_store}
\alias{tar_archive_store}
\title{Path to the archived target store directory}
\usage{
tar_archive_store(package, pipeline, store = targets::tar_config_get("store"))
}
\arguments{
\item{package}{A scalar character of the package name.}

\item{pipeline}{A scalar character of the pipeline name.}

\item{store}{Character of length 1, path to the
\code{targets} data store. Defaults to \code{tar_config_get("store")},
which in turn defaults to \verb{_targets/}.
When you set this argument, the value of \code{tar_config_get("store")}
is temporarily changed for the current function call.
See \code{\link[targets:tar_config_get]{tar_config_get()}} and \code{\link[targets:tar_config_set]{tar_config_set()}} for details
about how to set the data store path persistently
for a project.}
}
\value{
A scalar character of the path to the archived target store
directory.
}
\description{
Path to the archived target store directory
}
