% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_source_archive.R
\name{tar_source_archive}
\alias{tar_source_archive}
\title{Run archived R scripts.}
\usage{
tar_source_archive(
  package,
  files = "R",
  envir = targets::tar_option_get("envir"),
  change_directory = FALSE
)
}
\arguments{
\item{package}{A scalar character of the package name.}

\item{files}{Character vector of file and directory paths
to look for R scripts to run. Paths must either be absolute
paths or must be relative to the current working directory
just before the function call.}

\item{envir}{Environment to run the scripts. Defaults to
\code{tar_option_get("envir")}, the environment of the pipeline.}

\item{change_directory}{Logical, whether to temporarily change
the working directory to the directory of each R script
before running it.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Run archived R scripts.
}
