% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarchives-package.R
\docType{package}
\name{tarchives-package}
\alias{tarchives}
\alias{tarchives-package}
\title{tarchives: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
