% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tarchives.R
\name{use_tarchives}
\alias{use_tarchives}
\title{Use tarchives}
\usage{
use_tarchives(store = targets::tar_config_get("store"))
}
\arguments{
\item{store}{Character of length 1, path to the
\code{targets} data store. Defaults to \code{tar_config_get("store")},
which in turn defaults to \verb{_targets/}.
When you set this argument, the value of \code{tar_config_get("store")}
is temporarily changed for the current function call.
See \code{\link[targets:tar_config_get]{tar_config_get()}} and \code{\link[targets:tar_config_set]{tar_config_set()}} for details
about how to set the data store path persistently
for a project.}
}
\value{
No return value, called for side effects.
}
\description{
Set up tarchives for an existing package.
}
